refactor(daemon-service): finish the application surface split
`application_surface.rs` still carried ~2.7k lines of transport helpers
beside the wiring. Move them, unchanged, into the module tree:

- `registered_http`: the Work/Workflow/Handoff `RegisteredHttpOperation`
  impls, `registered_executor_unavailable`, `invoke_registered_http`
- `request_control`: the `application_http_context` middleware
- `operation_events` (new): the SSE operation-event and cancel routes
- `catalog` (new): the process-wide catalog snapshot and binding resolution
- `dispatch` (new): surface resolution, `execute_application_surface`,
  HTTP page projection and body parsing
- `feedback_observation` (new): argument-rejection observations
- `problems` (new): adapter/invocation problem envelopes

The root keeps the catalog dispatcher, invoker assembly, router assembly,
and re-exports every previously public name at its old path. Moved
root-private items become `pub(super)`, the same scope they had as
private items of the parent; tests import them from their new modules.
