+++ crates/tracedecay-daemon-service/src/application_surface/workflow.rs
@@ -14,7 +14,7 @@ use tracedecay_contracts::{
 };
 use tracedecay_tool_catalog::RouteExposureV1;
 
-use super::invoke_registered_http;
+use super::registered_http::invoke_registered_http;
 use tracedecay_daemon_protocol::ApplicationSurfaceAdapterError;
 use tracedecay_daemon_protocol::DaemonInvocationExecutor;
 use tracedecay_daemon_protocol::{WorkflowApplicationInvocation, WorkflowApplicationOutcome};
